Update code Signed-off-by: Arkadii Yakovets <2201626+arkid15r@users.noreply.github.com>
